% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSeries_callSubtypes2.R
\name{callEnsemble_Multi}
\alias{callEnsemble_Multi}
\title{callEnsemble_Multi}
\usage{
callEnsemble_Multi(
  X,
  ens = NULL,
  geneAnnotation = NULL,
  geneSet = NULL,
  scaller = NULL,
  geneid = "ensembl",
  subtype = c("PAD.train_20200110", "ImmuneSubtype")[1],
  verbose = T
)
}
\arguments{
\item{X}{a numeric RNA expression vector or matrix with gene names}

\item{ens}{A List of multiple model. Result of \code{\link{fitEnsembleModel}}}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneSet}{A list of genes for classification}

\item{scaller}{scaller object. A \code{\link[xgboost]{xgb.train}} object.}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}

\item{subtype}{Default subtype methods. Now, only \code{'PAD'} and
\code{'ImmuneSubtype'} are available. ATTENTION: If you use self-defined
data (ens, geneAnnoation, geneSet or scaller), you MUST set \code{subtype}
as \code{NULL}! All available options please visit
\code{\link{GSClassifier_Data}}}

\item{verbose}{whether report messages}
}
\value{
table, column 1 is best call, remaining columns are subtype
  prediction scores.
}
\description{
Make subtype calls for multiple samples
}
