% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_compareData.R
\name{compareData}
\alias{compareData}
\title{Compare data from the same source but different scale}
\usage{
compareData(
  X1,
  X2,
  names = c("X1", "X2"),
  width = 10,
  height = 10,
  savePath = "."
)
}
\arguments{
\item{X1}{Matrix with sample col and gene row}

\item{X2}{Matrix with sample col and gene row}

\item{names}{Label of X1 and X2}

\item{width, height}{Plot parameters}

\item{savePath}{Plot saved path}
}
\description{
Compare data from the same source but different scale
}
\examples{
## PAD genes
PAD <- readRDS(system.file("extdata", 'PAD.rds', package = "GSClassifier"))
tgene <- as.character(unlist(PAD$geneSet))
X1 <- stad_fpkm[tgene,]
X2 <- stad_tpm[tgene,]

## Compare FPKM with TPM matrix
res <- compareData(X1,X2,
                   names=c('FPKM','TPM'),
                   width = 10,
                   height = 10,
                   savePath = '.')
res2 <- compareData(log2(X1+1),log2(X2+1),
                    names=c('log2FPKM','log2TPM'),
                    width = 10,
                    height = 10,
                    savePath = '.')
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
