% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(vt, fromtype = "ENSEMBL", totype = "SYMBOL", db)
}
\arguments{
\item{vt}{Genes you want to annotation}

\item{fromtype}{The raw type of the genes.}

\item{totype}{The annotated type of the genes.}

\item{db}{A data frame containing annotation information.}
}
\description{
Multiple annotation based on dataset common.annot.Its colnames include:"ENSEMBL","ENSEMBL.Versions","SYMBOL","GENENAME","ENTREZID","UNIGENE" and "gene_type".
}
\details{
This function belongs to my private project call \code{lucky}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
