% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_createPtail.R
\name{createPtail}
\alias{createPtail}
\title{Greate Ptail Vector}
\usage{
createPtail(geneSet, interval = c(200, 400, 600, 800, 1000))
}
\arguments{
\item{geneSet}{List. Gene expression profiles - Co-expressed modules from WGCNA, etc}

\item{interval}{Numeric. Define the number of features you want in model training.}
}
\description{
Create Ptail Vector
}
\examples{
ImmuneSubtype <- readRDS(system.file("extdata", "ImmuneSubtype.rds", package = "GSClassifier"))
geneSet <- ImmuneSubtype$geneSet
ptail <- createPtail(geneSet, interval = c(200,400,600,800,1000))
print(ptail)
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
