% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_modelData.R
\name{modelData}
\alias{modelData}
\title{modelData}
\usage{
modelData(
  design,
  id.col = "ID",
  variable = c("platform", "PAD subtype"),
  Prop = 0.6,
  seed = 2020
)
}
\arguments{
\item{design}{A design object with ID rownames}

\item{id.col}{Colname of ID}

\item{variable}{Split factor for balanced selection}

\item{Prop}{the proportion of training cohort}

\item{seed}{random seed}
}
\description{
Training and internal validation data selection
}
