% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mv_tolerance.R
\name{mv_tolerance}
\alias{mv_tolerance}
\title{mv_tolerance}
\usage{
mv_tolerance(
  X,
  gene.loss = c(2, 4, 6, 8, 10, 12),
  levels = c(1, 2, 3, 4),
  model = "PAD.train_20220916",
  seed = 487,
  verbose = T
)
}
\arguments{
\item{X}{a data frame with the number of colume ≥ 2}

\item{gene.loss}{a vector of the number of genes masked}

\item{model}{a character of GSClassifier model or a \code{luckyModel} model}

\item{seed}{random seed}

\item{verbose}{whether to report the process}
}
\value{
A list containing results of subtype identification and multi-roc analysis
}
\description{
Missing value tolerance analysis for \code{GSClassifier} models
}
\examples{
Empty
}
\seealso{
\code{\link[pROC]{multiclass.roc}}.
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
