% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_PADi.R
\name{PADi}
\alias{PADi}
\title{PADi}
\usage{
PADi(
  X,
  geneid = "ensembl",
  cancer.type = "GC",
  version = c("20200110", "20220916")[1],
  numCores = 0,
  verbose = T
)
}
\arguments{
\item{X}{a numeric RNA expression vector or matrix with gene names}

\item{cancer.type}{The type of cancer}

\item{version}{model version}

\item{numCores}{No. of CPU core}

\item{verbose}{whether report messages}
}
\value{
A data frame containing personalized PAD subtypes
}
\description{
PAD for individuals
}
\details{
The current software only supports parameters: cancer.type = "GC", version = "20200110". More cancer types or model version were under developing.
}
\examples{
# Package
if (!requireNamespace("GSClassifier", quietly = TRUE)){
  devtools::install_github("huangwb8/GSClassifier")
}
library(GSClassifier)

# Load test data of RNA-Seq expression
testData <- readRDS(system.file("extdata", "testData.rds", package = "GSClassifier"))[['Kim2018_3']]
dim(testData) # 19118 3

# Data formats (Different data but the same function and usage)

# 1.Mutiple data
X = testData

# 2.Single data
# X <- testData[,1]; names(X) <- rownames(testData)
# X <- as.matrix(testData[,1]); rownames(X) <- rownames(testData)

## Method1: use a general function called 'callEnsemble'
res_padi <- callEnsemble(
  X = X,
  ens = NULL,
  geneAnnotation = NULL,
  geneSet = NULL,
  scaller = NULL,
  geneid = "ensembl",
  subtype = "PAD.train_20200110",
  verbose = F
)

## Method2: use a specific function called 'PADi'
res_padi <- PADi(X = X, verbose = F)

## Method3: Parallel strategy for lots of samples (empirically >50) to save #' time. Not Run for small cohorts.
# res_padi <- parCallEnsemble(
#   X = X,
#   ens = NULL,
#   geneAnnotation = NULL,
#   geneSet = NULL,
#   scaller = NULL,
#   geneids = 'ensembl',
#   subtype = 'PAD.train_20200110',
#   verbose = T,
#   numCores = 2)
# res_padi <- PADi(X = X, verbose = F, numCores = 4)
}
\seealso{
\code{\link{callEnsemble}}; \code{\link{parCallEnsemble}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
