% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{parCallEnsemble}
\alias{parCallEnsemble}
\title{parCallEnsemble}
\usage{
parCallEnsemble(
  X,
  ens = NULL,
  geneAnnotation = NULL,
  geneSet = NULL,
  scaller = NULL,
  geneids = "ensembl",
  subtype = c("PAD.train_20200110", "PAD.all_20200110", "ImmuneSubtype")[1],
  numCores = 2
)
}
\arguments{
\item{X}{gene expression matrix, genes in rows, samples in columns}

\item{ens}{A List of multiple model. Result of \code{\link{fitEnsembleModel}}}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneSet}{A list of genes for classification}

\item{scaller}{scaller object. A \code{\link[xgboost]{xgb.train}} object.}

\item{subtype}{Default subtype methods. Now, only \code{'PAD'} and
\code{'ImmuneSubtype'} are available. ATTENTION: If you use self-defined
data (ens, geneAnnoation, geneSet or scaller), you MUST set \code{subtype}
as \code{NULL}! All available options please visit
\code{\link{GSClassifier_Data}}}

\item{numCores}{No. of CPU core}
}
\value{
table, column 1 is best call, remaining columns are subtype
  prediction scores.
}
\description{
Parallel version: Make subtype calls for each sample
}
