% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/@luckyBase.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mymusic}
\title{#' @importFrom luckyBase convert
#' @export
luckyBase::convert
#' @importFrom luckyBase convert1
#' @export
luckyBase::convert1
#' @importFrom luckyBase convert2
#' @export
luckyBase::convert2
#' @importFrom luckyBase Plus.convert
#' @export
luckyBase::Plus.convert
#' @importFrom luckyBase classify
#' @export
luckyBase::classify
#' @importFrom luckyBase confint
#' @export
luckyBase::confint
#' @importFrom luckyBase cut_vector
#' @export
luckyBase::cut_vector
#' @importFrom luckyBase Fastextra
#' @export
luckyBase::Fastextra
#' @importFrom luckyBase Fastgrep
#' @export
luckyBase::Fastgrep
#' @importFrom luckyBase importFromPackage
#' @export
luckyBase::importFromPackage
#' @importFrom luckyBase is.one.na
#' @export
luckyBase::is.one.na
#' @importFrom luckyBase is.all.na
#' @export
luckyBase::is.all.na
#' @importFrom luckyBase is.one.ture
#' @export
luckyBase::is.one.ture
#' @importFrom luckyBase is.all.ture
#' @export
luckyBase::is.all.ture
#' @importFrom luckyBase LuckyVerbose
#' @export
luckyBase::LuckyVerbose
#' @importFrom luckyBase Plus.library
#' @export
luckyBase::Plus.library
#' @importFrom luckyBase show.color
#' @export
luckyBase::show.color
#' @importFrom luckyBase verbose_vector
#' @export
luckyBase::verbose_vector}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{luckyBase}{\code{\link[luckyBase]{mymusic}}}
}}

