% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_subtypeVector.R
\name{subtypeVector}
\alias{subtypeVector}
\title{subtypeVector}
\usage{
subtypeVector(
  expr,
  GEPs,
  cluster.method = c("ward.D2", "complete")[1],
  extra.annot = NULL,
  plot.title = NULL,
  k = NULL,
  outlier.filter = 2,
  verbose = T
)
}
\arguments{
\item{expr}{RNA expression matrix. Samples in col and ENSEMBL genes in row.}

\item{GEPs}{List. Gene expression profiles - Co-expressed modules from WGCNA, etc}

\item{cluster.method}{One of \code{'ward.D2'}, \code{'complete'} or other
methods in \code{\link[stats]{hclust}}. If \code{'randomForest'} was set,
the \code{\link[randomForest]{randomForest}} function would used to
classify the samples.}

\item{extra.annot}{Extra top annotation. The same order as colnames of \code{expr}.}

\item{plot.title}{The title of heatmap report.}

\item{k}{the number of clusters you want. Default is the same as the number of \code{GEPs} types}

\item{outlier.filter}{Integer. Set a small integer to remove outlier samples}

\item{verbose}{Whether to show heatmap in the process.}
}
\value{
A list: \cr
\itemize{
  \item \code{Repeat} Basic parameters
  \item \code{Data} mean expression of each GEP in each sample cluster
  \item \code{logloss} negative log-likelihood function
  \item \code{MissGene} Missing genes
  \item \code{Plot} a \code{\link[ComplexHeatmap]{Heatmap}} object
}
}
\description{
Automatic clustering based on expression matrix and gene expression profiles
}
\author{
Weibin Huang<\email{hwb2012@qq.com}>
}
