% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{trainDataProc}
\alias{trainDataProc}
\title{trainDataProc}
\usage{
trainDataProc(
  Xmat,
  Yvec,
  geneSet,
  subtype = 1,
  ptail = 0.5,
  breakVec = c(0, 0.25, 0.5, 0.75, 1)
)
}
\arguments{
\item{Xmat}{Matrix of gene expression, samples in columns, genes in rows}

\item{Yvec}{Vector of phenotype, strings, 1, 2, etc}

\item{geneSet}{A list of genes for classification}

\item{subtype}{cluster name, string '1', as in Yvec}

\item{ptail}{Binary phenotype vector.}

\item{breakVec}{vector of break points, used to bin expression data}
}
\value{
List of Xbin and Ybin, the binned, subset, and binarized values.
}
\description{
Binary data preprocessing based on expression matrix and real clustering
}
