% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSubtypes.R
\name{geneMatch}
\alias{geneMatch}
\title{geneMatch}
\usage{
geneMatch(X, geneAnnotation = NULL, geneid = "ensembl")
}
\arguments{
\item{X}{gene expression matrix, genes in rows, samples in columns}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneid}{type of gene id in \code{X}. One of \code{'symbol'}, \code{'entrez'} and \code{'ensembl'}.}
}
\value{
Matched index of genes to the expression matrix.
}
\description{
Match the incoming data to what was used in training
}
