% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSClassifier_Data.R
\name{GSClassifier_Data}
\alias{GSClassifier_Data}
\title{Data in \code{GSClassifier} package}
\usage{
GSClassifier_Data(ImmuneSubtype = NULL, PAD = NULL)
}
\arguments{
\item{ImmuneSubtype}{Data integration for calling of immune subtype by Thorsson et al. in 2018.}

\item{PAD}{Data integration for calling of pan-immune activation/dysfunction subtype (PAD).}
}
\description{
Internal dataset of GSClassifier package
}
\examples{
PAD <- readRDS(system.file("extdata", "PAD.train_20200110", package = "GSClassifier"))
ImmuneSubtype <- readRDS(system.file("extdata", "ImmuneSubtype.rds", package = "GSClassifier"))
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
