% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSeries_fitModels.R
\name{na_fill}
\alias{na_fill}
\title{NA filling with recursive partitioning and regression trees}
\usage{
na_fill(
  Xmat,
  method = c("quantile", "rpart", NULL)[1],
  seed = 2022,
  verbose = T
)
}
\arguments{
\item{Xmat}{A gene expression matrix with NA value}

\item{method}{One of \code{'quantile'}, \code{'rpart'} and \code{NULL}.}
}
\description{
NA filling with recursive partitioning and regression trees
}
\seealso{
\code{\link[rpart]{rpart}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
