% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSeries_callSubtypes2.R
\name{parCallEnsemble}
\alias{parCallEnsemble}
\title{parCallEnsemble}
\usage{
parCallEnsemble(
  X,
  ens = NULL,
  geneAnnotation = NULL,
  geneSet = NULL,
  scaller = NULL,
  geneids = "ensembl",
  matchmode = c("fix", "free")[1],
  subtype = c("PAD.train_20200110", "ImmuneSubtype")[1],
  verbose = T,
  numCores = 2
)
}
\arguments{
\item{X}{a numeric RNA expression vector or matrix with gene names}

\item{ens}{A List of multiple model. Result of \code{\link{fitEnsembleModel}}}

\item{geneAnnotation}{A data frame with ENSEMBL, SYMBOL and ENTREZID of the genes in \code{geneSet}}

\item{geneSet}{A list of genes for classification}

\item{scaller}{scaller object. A \code{\link[xgboost]{xgb.train}} object.}

\item{matchmode}{If your genes belong to ENSEMBL, SYMBOL or ENTREZ, you should set \code{matchmode = 'fix'}, in which all genes would be aligned to ENSEMBL. \code{'matchmode = free'} is a legacy feature, which means that you have to convert your genes into the same annotation type of the pre-trained \code{GSClassifier} model.}

\item{subtype}{Default subtype methods. Now, only \code{'PAD'} and
\code{'ImmuneSubtype'} are available. ATTENTION: If you use self-defined
data (ens, geneAnnoation, geneSet or scaller), you MUST set \code{subtype}
as \code{NULL}! All available options please visit
\code{\link{GSClassifier_Data}}}

\item{verbose}{whether report messages}

\item{numCores}{No. of CPU core}
}
\value{
table, column 1 is best call, remaining columns are subtype
  prediction scores.
}
\description{
Parallel version of callEnsemble
}
\details{
1. Data of one sample was not supported. Please use \code{\link{callEnsemble}} instead. \cr
2. The sample of \code{X} must be over 10!
}
