% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/@environment.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{convert}
\alias{LuckyVerbose}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{luckyBase}{\code{\link[luckyBase]{LuckyVerbose}}, \code{\link[luckyBase]{convert}}}
}}

